Move Ability definition to web-client
