+++ packages/web-app-admin-settings/src/index.ts
@@ -3,7 +3,7 @@ import General from './views/General.vue'
 import Users from './views/Users.vue'
 import Groups from './views/Groups.vue'
 import Spaces from './views/Spaces.vue'
-import { Ability } from 'web-pkg'
+import { Ability } from 'web-client/src/helpers/resource/types'
 import { AppNavigationItem } from 'web-pkg/src/apps'
 
 // just a dummy function to trick gettext tools
