+++ packages/web-app-files/tests/unit/helpers/share/link.spec.ts
@@ -3,7 +3,7 @@ import { DateTime } from 'luxon'
 import { Store } from 'vuex'
 import { ClientService } from 'web-pkg/src/services'
 import { useClipboard } from '@vueuse/core'
-import { Ability } from 'web-pkg'
+import { Ability } from 'web-client/src/helpers/resource/types'
 import { mock, mockDeep } from 'jest-mock-extended'
 import { Language } from 'vue3-gettext'
 import { Resource } from 'web-client'
