+++ packages/web-client/package.json
@@ -16,6 +16,7 @@
     "generate-openapi": "rm -rf src/generated && docker run --rm -v \"${PWD}/src:/local\" openapitools/openapi-generator-cli generate -i https://raw.githubusercontent.com/owncloud/libre-graph-api/main/api/openapi-spec/v1.0.yaml  -g typescript-axios -o /local/generated"
   },
   "peerDependencies": {
+    "@casl/ability": "^6.3.3",
     "axios": "^0.27.2",
     "lodash-es": "^4.17.21"
   }
