+++ packages/web-pkg/package.json
@@ -14,7 +14,6 @@
     "directory": "packages/web-pkg"
   },
   "peerDependencies": {
-    "@casl/ability": "^6.3.3",
     "@casl/vue": "^2.2.1",
     "axios": "^0.27.2",
     "filesize": "^9.0.11",
