+++ packages/web-pkg/src/composables/ability/useAbility.ts
@@ -1,4 +1,4 @@
 import { useAbility as _useAbility } from '@casl/vue'
-import { Ability } from './../../utils'
+import { Ability } from 'web-client/src/helpers/resource/types'
 
 export const useAbility = () => _useAbility<Ability>()
