+++ packages/web-test-helpers/src/defaultPlugins.ts
@@ -3,7 +3,7 @@ import { createGettext } from 'vue3-gettext'
 import { h } from 'vue'
 import { abilitiesPlugin } from '@casl/vue'
 import { createMongoAbility } from '@casl/ability'
-import { AbilityRule } from 'web-pkg/src'
+import { AbilityRule } from 'web-client/src/helpers/resource/types'
 
 export interface DefaultPluginsOptions {
   designSystem?: boolean
